Remove old SBJsonParser and rename SBJsonChunkParser to SBJsonParser
- Updated tests to suit.
- Rename item "return" block.
- Rename the main constructor / initialiser. (Again.)
- Add two more constructor methods for the parser.
