+++ src/main/objc/SBJson.h
@@ -27,10 +27,9 @@
  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#import "SBJsonParser.h"
 #import "SBJsonWriter.h"
 #import "SBJsonStreamParser.h"
-#import "SBJsonChunkParser.h"
+#import "SBJsonParser.h"
 #import "SBJsonStreamWriter.h"
 #import "SBJsonStreamTokeniser.h"
